requirements.txt
pandas
openai
chromadb
flask
python-dotenv
sentence-transformers
pytest